Finshed most change requests
