+++ .gitignore
@@ -1,3 +1,4 @@
-/target
+target/
+.idea/
 **/*.rs.bk
 Cargo.lock
