+++ src/lib.rs
@@ -9,5 +9,5 @@ extern crate serde_derive;
 mod net;
 mod packet;
 
-use packet::Packet;
+use packet::{Packet, RawPacket};
 pub use net::UdpSocket;
