+++ src/net/local_ack.rs
@@ -1,5 +1,5 @@
+use Packet;
 use std::collections::HashMap;
-use packet::Packet;
 
 /// Packets waiting for an ack
 ///
@@ -7,24 +7,27 @@ use packet::Packet;
 ///
 /// Additionally, holds packets "forward" of the current ack packet
 #[derive(Debug)]
-pub struct AckRecord {
+pub struct LocalAckRecord {
+    // packets waiting for acknowledgement.
     packets: HashMap<u16, Packet>
 }
 
-impl AckRecord {
-    pub fn new() -> AckRecord {
-        AckRecord { packets: HashMap::new() }
+impl LocalAckRecord {
+    pub fn new() -> LocalAckRecord {
+        LocalAckRecord { packets: HashMap::new() }
     }
 
+    /// Checks if there are packets in the queue to be aknowleged.
     pub fn is_empty(&mut self) -> bool {
         self.packets.is_empty()
     }
 
+    /// Gets the total packages in the queue that could be aknowleged.
     pub fn len(&mut self) -> usize {
         self.packets.len()
     }
 
-    /// Adds a packet to the waiting packets
+    /// Adds a packet to the queue awaiting for an aknowlegement.
     pub fn enqueue(&mut self, seq: u16, packet: Packet) {
         // TODO: Handle overwriting other packet?
         //   That really shouldn't happen, but it should be encoded here
@@ -56,48 +59,5 @@ impl AckRecord {
         }
 
         dropped_packets.into_iter().map(|seq| (seq, self.packets.remove(&seq).unwrap())).collect()
-    }
-}
-
-/// Third party's ack information
-///
-/// Holds the latest seq_num we've seen from them and the 32 bit bitfield
-/// for extra redundancy
-#[derive(Debug)]
-pub struct ExternalAcks {
-    pub last_seq: u16,
-    pub field: u32,
-    initialized: bool
-}
-
-impl ExternalAcks {
-    pub fn new() -> ExternalAcks {
-        ExternalAcks { last_seq: 0, field: 0, initialized: false }
-    }
-
-    pub fn ack(&mut self, seq_num: u16) {
-        if !self.initialized {
-            self.last_seq = seq_num;
-            self.initialized = true;
-            return;
-        }
-
-        let pos_diff = seq_num.wrapping_sub(self.last_seq);
-        let neg_diff = self.last_seq.wrapping_sub(seq_num);
-
-        if pos_diff == 0 {
-            return;
-        }
-
-        if pos_diff < 32000 {
-            if pos_diff <= 32 {
-                self.field = ((self.field << 1 ) | 1) << (pos_diff - 1);
-            } else {
-                self.field = 0;
-            }
-            self.last_seq = seq_num;
-        } else if neg_diff <= 32 {
-            self.field = self.field | (1 << neg_diff - 1);
-        }
     }
 }
