+++ src/net/mod.rs
@@ -1,9 +1,14 @@
 mod connection;
-mod packet;
 mod udp;
+mod external_ack;
+mod local_ack;
+mod socket_state;
 
-use self::connection::{Connection, ConnectionQuality };
-use self::packet::{ AckRecord, ExternalAcks };
-use Packet;
+use std::net::SocketAddr;
+pub use self::connection::{Connection, ConnectionQuality };
+use self::local_ack::LocalAckRecord;
+use self::external_ack::ExternalAcks;
+use self::socket_state::SocketState;
+use super::{Packet, RawPacket};
 
 pub use self::udp::UdpSocket;
